:sparkles: (urls) unfurl urls on bullet publishing

+++ app/Http/Controllers/BulletController.php
@@ -7,6 +7,7 @@ use App\Models\Bullet;
 use Exception;
 use Givebutter\LaravelKeyable\Auth\AuthorizesKeyableRequests;
 use Illuminate\Http\JsonResponse;
+use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 
@@ -15,16 +16,38 @@ class BulletController extends Controller
     use AuthorizesKeyableRequests;
 
     /**
+     * @param Request $request
+     * @return JsonResponse
+     */
+    public function index(Request $request)
+    {
+        $user = request()->keyable;
+
+        if (!is_null($user)) {
+            return JsonResponse::create([
+                'message' => '📓 Here is your entire journal.',
+                'data' => $user->bullets->groupBy(function ($item) {
+                    return $item->published_at->format('d-m-y');
+                })
+            ]);
+        } else {
+            return JsonResponse::create([
+                'message' => '❌ You need an API key to use this endpoint.'
+            ], 400);
+        }
+    }
+
+    /**
      * Store a newly created bullet in storage.
      *
      * @param Request $request
-     * @return JsonResponse|Response
+     * @return JsonResponse|RedirectResponse|Response
      */
     public function store(Request $request)
     {
         $bullet = null;
         $message = null;
-        $user = is_null($request->user()) ? $request->keyable : $request->user();
+        $user = is_null($request->user()) ? request()->keyable : $request->user();
         $messages = [
             'success' => '✅ Bullet saved!',
             'error' => '❌ Could not save bullet.'
@@ -34,7 +57,8 @@ class BulletController extends Controller
             $bullet = Bullet::create([
                 'user_id' => $user->id,
                 'published_at' => now()->timezone($user->timezone),
-                'bullet' => $request->bullet
+                'bullet' => $request->bullet,
+                'urls' => unfurl_string($request->bullet)
             ]);
 
             if (!$request->wantsJson() && $request->hasSession()) {
@@ -51,7 +75,7 @@ class BulletController extends Controller
         }
 
         if (!$request->wantsJson() && $request->hasSession()) {
-            return redirect('journal');
+            return redirect()->to('journal');
         } else {
             return JsonResponse::create([
                 'message' => $message,
@@ -63,22 +87,50 @@ class BulletController extends Controller
     /**
      * Destroy a specified bullet.
      *
+     * @param Request $request
      * @param $id
-     * @return Response
+     * @return JsonResponse|RedirectResponse|Response
      */
-    public function delete($id)
+    public function delete(Request $request, $id)
     {
+        $bullet = null;
+        $message = null;
+        $user = is_null($request->user()) ? request()->keyable : $request->user();
+        $messages = [
+            'success' => '✅ Bullet deleted!',
+            'error' => '❌ Could not delete bullet.'
+        ];
+
         try {
-            $bullet = Bullet::find($id);
+            $bullet = $user->bullets->find($id);
+
+            if (is_null($bullet)) {
+                throw new Exception('bullet_not_found');
+            }
 
             $bullet->delete();
 
-            request()->session()->flash('success', '✅ Bullet deleted!');
+            if (!$request->wantsJson() && $request->hasSession()) {
+                $request->session()->flash('success', $messages['success']);
+            } else {
+                $message = $messages['success'];
+            }
         } catch (Exception $e) {
-            request()->session()->flash('error', '❌ Could not delete bullet.');
+            if (!$request->wantsJson() && $request->hasSession()) {
+                $request->session()->flash('error', $messages['error']);
+            } else {
+                $message = $messages['error'];
+            }
         }
 
-        return redirect('journal');
+        if (!$request->wantsJson() && $request->hasSession()) {
+            return redirect()->to('journal');
+        } else {
+            return JsonResponse::create([
+                'message' => $message,
+                'data' => null
+            ]);
+        }
     }
 
     /**
@@ -86,22 +138,49 @@ class BulletController extends Controller
      *
      * @param BulletUpdateRequest $request
      * @param $id
-     * @return Response
+     * @return JsonResponse|RedirectResponse|Response
      */
     public function update(BulletUpdateRequest $request, $id)
     {
+        $bullet = null;
+        $message = null;
+        $user = is_null($request->user()) ? request()->keyable : $request->user();
+        $messages = [
+            'success' => '✅ Bullet updated!',
+            'error' => '❌ Could not update bullet.'
+        ];
+
         try {
-            $bullet = Bullet::find($id);
+            $bullet = $user->bullets->find($id);
+
+            if (is_null($bullet)) {
+                throw new Exception('bullet_not_found');
+            }
 
             $bullet->bullet = $request->bullet;
 
             $bullet->save();
 
-            request()->session()->flash('success', '✅ Bullet updated!');
+            if (!$request->wantsJson() && $request->hasSession()) {
+                $request->session()->flash('success', $messages['success']);
+            } else {
+                $message = $messages['success'];
+            }
         } catch (Exception $e) {
-            request()->session()->flash('error', '❌ Could not update bullet.');
+            if (!$request->wantsJson() && $request->hasSession()) {
+                $request->session()->flash('error', $messages['error']);
+            } else {
+                $message = $messages['error'];
+            }
         }
 
-        return redirect('journal');
+        if (!$request->wantsJson() && $request->hasSession()) {
+            return redirect()->to('journal');
+        } else {
+            return JsonResponse::create([
+                'message' => $message,
+                'data' => $bullet
+            ]);
+        }
     }
 }
